+++ app-layout-ui-tests/app-layout-ui-test-top-large/src/main/java/com/github/appreciated/app/layout/test/top/large/view/View3.java
@@ -1,12 +1,12 @@
-package com.github.appreciated.app.layout.test.toplarge.view;
+package com.github.appreciated.app.layout.test.top.large.view;
 
-import com.github.appreciated.app.layout.test.toplarge.TopLargeBehaviourView;
+import com.github.appreciated.app.layout.test.top.large.TopLargeBehaviourView;
 import com.github.appreciated.app.layout.test.view.ExampleView;
 import com.vaadin.flow.router.Route;
 
-@Route(value = "view8", layout = TopLargeBehaviourView.class)
+@Route(value = "view3", layout = TopLargeBehaviourView.class)
 // an empty view name will also be the default view
-public class View8 extends ExampleView {
+public class View3 extends ExampleView {
     @Override
     protected String getViewName() {
         return getClass().getName();
